requirements.txt
pandas
numpy
pandas-gbq
google-cloud-bigquery
google-cloud-bigquery-storage
hdbscan
toml
google-cloud-pubsub
scikit-learn==1.5.2